Fix Age Restricted Videos Without Sign In
I don't know why these videos exist, but sometimes there are videos which label
themselves as being age restricted, but don't require the user to sign in to
verify their age.

These were causing things to fail, since the embedPage didn't exist, since we
never needed to go through that code path.
The good news, though, is that we didn't need the embedPage because we had all
the info we needed on the current page.

We just had to reach out and use it.