Remove PlayerScript Requirement
So... this exists.
Apparently there are videos which are protected, but don't have an "s" parameter
and don't need signatures.

Essentially, it tries "embedded" tag, but that fails because it's protected, so
it tries "display_page".

That fails because it's age restricted.

We then do into the "no tags left" part of startNextRequest.
That loads the web, which loads the embed, which loads the javascript, which
finally loads the video_info.

We now have a playerScript, but don't need to sign requests.
We never returned the CipherRequired, we just went through a path that loaded
the JS anyway.

Neat stuff.