More Restrictive Error Case
Turns out there was a reason for that to fail.
If the signature should exist and doesn't, then it's a real failure.

So, in this case I've changed it to only fail if there's a playerScript without
a signature, where there was something to deescramble in the first place.
Seems to solve both issues.